refact: remove namelist types from columnsetmanager
